Complete restructure translate_parameter_constrain
This commit completes the restructuring of the parameter table
generation. The parameter table now misses file expressions and all
contraint comands are individual columns in the table. A column with the
initial full parameter name was added. This commit requires the
following changes in other functions:

- Generating filter commands when reading initial SWAT2012 parameter
- Generating filter commands for parameter changes.

Missing changes in translate_parameter_constraint

- routine to check the constraint variables for SWAT2012 and +